docs: add stream examples
