+++ example_test.go
@@ -155,7 +155,28 @@ func ExampleClient_ConnectToStream() {
 	fmt.Println("done")
 }
 func ExampleClient_VolumeStreams() {
-
+	client := gotwtr.New("key")
+	ch := make(chan gotwtr.VolumeStreamsResponse, 5)
+	errCh := make(chan error)
+	stream := client.VolumeStreams(context.Background(), ch, errCh)
+	fmt.Println("streaming...")
+	done := make(chan struct{})
+	go func(done chan struct{}) {
+		for {
+			select {
+			case data := <-ch:
+				fmt.Println(data.Tweet)
+			case err := <-errCh:
+				fmt.Println(err)
+			case <-done:
+				break
+			}
+		}
+	}(done)
+	time.Sleep(time.Second * 10)
+	close(done)
+	stream.Stop()
+	fmt.Println("done")
 }
 func ExampleClient_RetweetsLookup() {
